# backend/openapi/openapi-1.18.0-beta/openapi-1.18.0-beta.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cyber Media Center API (赛博影视)",
    "description": "## 专业发烧级影视中心后端接口文档\n\n此文档定义了 Cyber Media 的核心 API 交互规范。旨在取代 Plex/Emby，提供高性能、低延迟的媒体管理与流媒体服务。\n\n### 🔗 API 生命周期说明 (x-lifecycle-status)\n我们在每个接口中定义了开发状态，请前端同学注意：\n\n1.  🔴 **pending** (未开发): 接口仅有定义，后端尚未实现，调用会返回 501。\n2.  🟠 **implemented_unverified** (已开发未验证): 代码已合并，但未经完整测试，可能存在 Bug。\n3.  ❌ **failed_verification** (验证失败): 测试发现严重问题，暂时禁止使用。\n4.  ⚠️ **passed_imperfect** (验证通过但格式不规范): 功能可用，但返回字段可能不全或命名不符合规范，后续会变动。\n5.  ✅ **stable** (验证通过): 接口稳定，可放心集成。\n\n### 🛡️ 鉴权说明\n所有非公开接口需在 Header 中携带 `Authorization: Bearer <token>`。",
    "version": "1.18.0-beta",
    "contact": {
      "name": "Cyber Media Core Team",
      "email": "dev@cybermedia.local"
    }
  },
  "servers": [
    {
      "url": "https://pw.pioneer.fan:84",
      "description": "生产环境 (Production HTTPS)"
    },
    {
      "url": "http://pioneer.fan:84",
      "description": "生产环境 (Production HTTP IPv4)"
    },
    {
      "url": "http://localhost:5004",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/api/v1/storage/preview": {
      "post": {
        "summary": "✅ [验证通过] 预览/浏览存储源目录",
        "description": "无需保存来源即可验证配置并浏览目录。当前支持 local、webdav、smb、ftp、alist、openlist。返回 browse payload，供前端目录选择器直接使用。",
        "operationId": "previewStorage",
        "x-lifecycle-status": "implemented_unverified",
        "tags": [
          "Storage"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoragePreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "目录浏览结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoragePreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "连接失败或配置错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/storage/sources": {
      "get": {
        "summary": "✅ [验证通过] 获取存储源列表",
        "description": "列出所有已配置的存储源。",
        "operationId": "listStorageSources",
        "x-lifecycle-status": "stable",
        "tags": [
          "Storage"
        ],
        "responses": {
          "200": {
            "description": "存储源列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StorageSource"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "🟠 [已开发未验证] 添加存储源",
        "description": "添加新的媒体库来源。",
        "operationId": "addStorageSource",
        "x-lifecycle-status": "implemented_unverified",
        "tags": [
          "Storage"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storage/sources/{id}": {
      "delete": {
        "summary": "🟠 [已开发未验证] 删除存储源",
        "description": "移除指定的存储源配置。",
        "operationId": "deleteStorageSource",
        "x-lifecycle-status": "implemented_unverified",
        "tags": [
          "Storage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "存储源 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keep_metadata",
            "in": "query",
            "description": "是否保留元数据。",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "无法删除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "未找到指定存储源"
          }
        }
      },
      "patch": {
        "summary": "🟠 [已开发未验证] 更新存储源配置",
        "description": "更新存储源的连接信息或别名。",
        "operationId": "updateStorageSource",
        "x-lifecycle-status": "implemented_unverified",
        "tags": [
          "Storage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "存储源 ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "✅ [验证通过] 获取单个存储源详情",
        "description": "返回单个存储源的脱敏配置、能力和前端可用动作。",
        "operationId": "getStorageSource",
        "x-lifecycle-status": "stable",
        "tags": [
          "Storage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "存储源 ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "存储源详情",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StorageSource"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "存储源不存在"
          }
        }
      }
    },
    "/api/v1/storage/sources/{id}/browse": {
      "get": {
        "summary": "✅ [验证通过] 浏览已保存存储源目录",
        "description": "基于已保存的存储源配置浏览目录，主要用于资源库绑定时选择 root_path。支持 local/webdav/smb/ftp/alist/openlist。",
        "operationId": "browseStorageSource",
        "x-lifecycle-status": "stable",
        "tags": [
          "Storage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "存储源 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "相对来源根目录的目标目录",
            "schema": {
              "type": "string",
              "default": "/"
            }
          },
          {
            "name": "dirs_only",
            "in": "query",
            "description": "是否只返回目录项，默认 true",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "目录浏览结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageSourceBrowseResponse"
                }
              }
            }
          },
          "400": {
            "description": "路径无效、连接失败或查询参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "存储源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/storage/sources/{id}/scan": {
      "post": {
        "summary": "✅ [验证通过] 扫描指定存储源",
        "description": "仅对指定的存储源触发扫描。支持通过 root_path 或 target_path 将扫描范围限定到预览选中的子目录；未传时按存储源根目录扫描。",
        "operationId": "scanSource",
        "x-lifecycle-status": "stable",
        "tags": [
          "Storage"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageSourceScanRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "存储源 ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "扫描任务已接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "msg": {
            "type": "string",
            "example": "success"
          },
          "trace_id": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "msg": {
            "type": "string"
          }
        }
      },
...
      "StoragePreviewRequest": {
        "type": "object",
        "description": "存储源临时预览请求，不会写入数据库；支持 local/webdav/smb/ftp/alist/openlist。",
        "required": [
          "type",
          "config"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "local",
              "webdav",
              "smb",
              "ftp",
              "alist",
              "openlist"
            ],
            "description": "存储协议类型"
          },
          "config": {
            "description": "连接配置，按 type 选择对应 schema。",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfigLocal"
              },
              {
                "$ref": "#/components/schemas/ConfigWebDAV"
              },
              {
                "$ref": "#/components/schemas/ConfigSMB"
              },
              {
                "$ref": "#/components/schemas/ConfigFTP"
              },
              {
                "$ref": "#/components/schemas/ConfigAList"
              },
              {
                "$ref": "#/components/schemas/ConfigOpenList"
              }
            ]
          },
          "target_path": {
            "type": "string",
            "default": "/",
            "description": "相对来源根目录的目标目录"
          },
          "dirs_only": {
            "type": "boolean",
            "default": true,
            "description": "是否只返回目录项，默认 true"
          }
        }
      },
      "StorageSource": {
        "type": "object",
        "description": "存储源配置实体",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "WEBDAV"
          },
          "type": {
            "type": "string",
            "enum": [
              "local",
              "webdav",
              "smb",
              "ftp",
              "alist",
              "openlist"
            ],
            "example": "webdav"
          },
          "root_path": {
            "type": "string",
            "description": "显示给前端的路径摘要"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "unknown",
              "unsupported"
            ]
          },
          "display_name": {
            "type": "string",
            "description": "协议显示名",
            "example": "WebDAV"
          },
          "is_supported": {
            "type": "boolean",
            "description": "当前后端是否真的支持该协议",
            "example": true
          },
          "capabilities": {
            "$ref": "#/components/schemas/StorageProviderCapabilities"
          },
          "config": {
            "type": "object",
            "description": "脱敏后的当前配置摘要",
            "additionalProperties": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfigLocal"
              },
              {
                "$ref": "#/components/schemas/ConfigWebDAV"
              },
              {
                "$ref": "#/components/schemas/ConfigSMB"
              },
              {
                "$ref": "#/components/schemas/ConfigFTP"
              },
              {
                "$ref": "#/components/schemas/ConfigAList"
              },
              {
                "$ref": "#/components/schemas/ConfigOpenList"
              }
            ]
          },
          "actions": {
            "$ref": "#/components/schemas/StorageSourceActions"
          },
          "health": {
            "$ref": "#/components/schemas/StorageSourceHealth"
          },
          "config_valid": {
            "type": "boolean",
            "description": "当前配置是否能通过后端静态校验"
          },
          "config_error": {
            "type": "string",
            "nullable": true,
            "description": "配置校验错误信息，合法时为 null"
          },
          "usage": {
            "$ref": "#/components/schemas/StorageSourceUsage"
          },
          "guards": {
            "$ref": "#/components/schemas/StorageSourceGuards"
          }
        }
      },
      "StorageSourceRequest": {
        "type": "object",
        "description": "添加或更新存储源请求体",
        "required": [
          "name",
          "type",
          "config"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "My Cloud Disk",
            "description": "自定义别名"
          },
          "type": {
            "type": "string",
            "enum": [
              "local",
              "webdav",
              "smb",
              "ftp",
              "alist",
              "openlist"
            ],
            "description": "存储协议类型"
          },
          "config": {
            "description": "根据协议类型动态变化的配置对象",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConfigLocal"
              },
              {
                "$ref": "#/components/schemas/ConfigWebDAV"
              },
              {
                "$ref": "#/components/schemas/ConfigSMB"
              },
              {
                "$ref": "#/components/schemas/ConfigFTP"
              },
              {
                "$ref": "#/components/schemas/ConfigAList"
              },
              {
                "$ref": "#/components/schemas/ConfigOpenList"
              }
            ]
          }
        }
      },
      "ConfigLocal": {
        "type": "object",
        "title": "Local Config",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string",
            "description": "服务器本地绝对路径",
            "example": "/data/media/movies"
          }
        }
      },
      "ConfigWebDAV": {
        "type": "object",
        "title": "WebDAV Config",
        "required": [
          "host",
          "username",
          "password"
        ],
        "properties": {
          "host": {
            "type": "string",
            "description": "域名或IP (不含 http 前缀)"
          },
          "port": {
            "type": "integer",
            "default": 443
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "secure": {
            "type": "boolean",
            "default": true,
            "description": "True=HTTPS, False=HTTP"
          },
          "root": {
            "type": "string",
            "default": "/",
            "description": "WebDAV 挂载根路径"
          }
        }
      },
...
      "StorageSourceScanRequest": {
        "type": "object",
        "properties": {
          "root_path": {
            "type": "string",
            "description": "限定扫描起始目录，通常复用预览选中的目录。",
            "example": "电影/华语"
          },
          "target_path": {
            "type": "string",
            "description": "root_path 的兼容别名；两者同时存在时优先使用 root_path。",
            "example": "/电影/华语"
          },
          "content_type": {
            "type": "string",
            "nullable": true,
            "description": "可选内容类型提示，用于优化电影/剧集识别。",
            "example": "movie"
          },
          "scrape_enabled": {
            "type": "boolean",
            "default": true,
            "description": "是否在扫描完成后继续触发刮削。"
          },
          "scraper_policy": {
            "$ref": "#/components/schemas/ScraperPolicy"
          },
          "provider_order": {
            "type": "array",
            "description": "scraper_policy.provider_order 的扁平别名。",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "StoragePreviewData": {
        "type": "object",
        "properties": {
          "storage_type": {
            "type": "string",
            "enum": [
              "local",
              "webdav",
              "smb",
              "ftp",
              "alist",
              "openlist"
            ]
          },
          "current_path": {
            "type": "string",
            "example": "/"
          },
          "parent_path": {
            "type": "string",
            "nullable": true,
            "example": "/"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StorageBrowseItem"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/StorageProviderCapabilities"
          }
        }
      },
      "StoragePreviewResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/StoragePreviewData"
              }
            }
          }
        ]
      },
...
      "ConfigOpenList": {
        "type": "object",
        "title": "OpenList Config",
        "description": "OpenList 当前按 AList 兼容 API 接入。",
        "allOf": [
          {
            "$ref": "#/components/schemas/ConfigAList"
          }
        ]
      },
...
      "StorageSourceBrowseResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/StorageSourceBrowseData"
              }
            }
          }
        ]
      },
...
    }
  }
}